endpoint: Refactor identifiers to separate files
To assist the sharing of code and appropriate locking access, move this
code around such that there is a dedicated endpointid.Identifiers type
to collect Identifiers for endpoint, such as the Cilium Endpoint ID,
container ID, k8s namespace/name, etc. Furthermore, move the access and
handling of these types all to a single file so they're in one place
rather than spread across multiple files.
